+++ packages/jetbrains-plugin/src/test/resources/project-fixtures/basic-java-project-with-mongodb/src/main/java/alt/mongodb/javadriver/JavaDriverRepository.java
@@ -42,4 +42,14 @@ public class JavaDriverRepository {
             ).sort(Sorts.ascending("year"))
             .first();
     }
+
+    public Document notIndexedAnotherQuery(Language lang) {
+        return client
+            .getDatabase("sample_mflix")
+            .getCollection("movies")
+            .find(
+                Filters.eq("rated", Rate.TV_PG)
+            )
+            .first();
+    }
 }