+++ packages/mongodb-dialects/java-driver/src/main/kotlin/com/mongodb/jbplugin/dialects/javadriver/glossary/JavaDriverDialectFormatter.kt
@@ -2,14 +2,37 @@ package com.mongodb.jbplugin.dialects.javadriver.glossary
 
 import com.mongodb.jbplugin.dialects.DialectFormatter
 import com.mongodb.jbplugin.dialects.OutputQuery
-import com.mongodb.jbplugin.mql.*
+import com.mongodb.jbplugin.indexing.IndexAnalyzer
+import com.mongodb.jbplugin.mql.BsonAny
+import com.mongodb.jbplugin.mql.BsonAnyOf
+import com.mongodb.jbplugin.mql.BsonArray
+import com.mongodb.jbplugin.mql.BsonBoolean
+import com.mongodb.jbplugin.mql.BsonDate
+import com.mongodb.jbplugin.mql.BsonDecimal128
+import com.mongodb.jbplugin.mql.BsonDouble
+import com.mongodb.jbplugin.mql.BsonEnum
+import com.mongodb.jbplugin.mql.BsonInt32
+import com.mongodb.jbplugin.mql.BsonInt64
+import com.mongodb.jbplugin.mql.BsonNull
+import com.mongodb.jbplugin.mql.BsonObject
+import com.mongodb.jbplugin.mql.BsonObjectId
+import com.mongodb.jbplugin.mql.BsonString
+import com.mongodb.jbplugin.mql.BsonType
+import com.mongodb.jbplugin.mql.BsonUUID
+import com.mongodb.jbplugin.mql.Node
+import com.mongodb.jbplugin.mql.QueryContext
 
 object JavaDriverDialectFormatter : DialectFormatter {
     override suspend fun <S> formatQuery(query: Node<S>, queryContext: QueryContext) =
         OutputQuery.None
 
-    override fun <S> indexCommandForQuery(query: Node<S>) =
+    override fun <S> indexCommand(
+        query: Node<S>,
+        index: IndexAnalyzer.SuggestedIndex<S>,
+        toQueryReference: (Node<S>) -> String?
+    ): String {
         throw UnsupportedOperationException()
+    }
 
     override fun formatType(type: BsonType): String = when (type) {
         is BsonDouble -> "double"